Tests run for Django 2
