+++ setup.py
@@ -5,14 +5,14 @@ setup(
     name='django-synchro',
     description='Django app for database data synchronization.',
     long_description=open('README.rst').read(),
-    version='0.7',
+    version='0.8',
     author='Jacek Tomaszewski',
     author_email='jacek.tomek@gmail.com',
     url='https://github.com/zlorf/django-synchro',
     license='MIT',
     install_requires=(
-        'django-dbsettings>=0.7',
-        'django>=1.7',
+        'django-dbsettings>=0.11',
+        'django>=2.2.0',
     ),
     classifiers=[
         'Development Status :: 4 - Beta',
@@ -21,11 +21,7 @@ setup(
         'Operating System :: OS Independent',
         'Programming Language :: Python',
         'Framework :: Django',
-        'Framework :: Django :: 1.7',
-        'Framework :: Django :: 1.8',
-        'Framework :: Django :: 1.9',
-        'Framework :: Django :: 1.10',
-        'Framework :: Django :: 1.11',
+        'Framework :: Django :: 2.2',
     ],
     packages=find_packages(),
     include_package_data = True,
