+++ synchro/urls.py
@@ -1,9 +1,10 @@
 # flake8: noqa
 from django.conf.urls import url
+from django.urls import path
 
 from .views import synchro
 
 
 urlpatterns = (
-    url(r'^$', synchro, name='synchro'),
+    path('', synchro, name='synchro'),
 )
