+++ tests/test_urls.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from django.contrib import admin
 from django.conf.urls import url, include
